requirements.txt
fastapi
uvicorn
litellm
python-dotenv
python-telegram-bot
pydantic